app/requirements.txt
anyio==3.6.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bech32==1.2.0
boto3==1.17.58
botocore==1.20.58
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
crispy-bootstrap5==0.6
cryptography==36.0.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
Django==4.0.1
django-bootstrap5==21.3
django-crispy==1.8.1
django-crispy-forms==1.13.0
django-email-verification==0.3.1
django-environ==0.9.0
django-js-asset==2.0.0
django-quill-editor==0.1.40
django-recaptcha==3.0.0
django-storages==1.11.1
fastapi==0.78.0
filelock==3.4.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
image==1.5.33
jmespath==0.10.0
oauthlib==3.2.2
packaging==20.9
piexif==1.1.3
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
PyYAML==6.0
qrcode==6.1
requests==2.27.1
requests-oauthlib==1.3.0
s3transfer==0.4.0
segno==1.5.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.8
uvicorn==0.18.2
uvloop==0.16.0
virtualenv==20.10.0
watchfiles==0.15.0
websockets==10.3
